Updates to misfit and HMC propagation functions. Updated the gradient of the misfit into an analytic representation of any linear system, with the misfit functional composed of data and prior knowledge. This includes precomputing some large matrices which are stored in the subclasses of the montecarlo class.
After a lot of trial and error, it appeared that the proposed momentum
was off by a large factor. It turned out that one should not use the
inverse of the inverse covariance matrix. All seems to work well now,
and trajectories oscillate through model (phase) space.

An extra piece of python was written to examine the gradient of the
misfit function. Typical values for standard deviation in measurement
error and prior information (typically on the order of 10-50%) seem to
produce very typical gradients; centered near the actual input
parameters of the synthetic data. The offset is caused by the prior
information. It does not have its own C++ subprogram, but instead is
present in runSampling. Comment and uncomment code as seems fit.