Added reading linear systems from files, augmented MC sampling. Added reading linear systems and data in a clear way from files, which simplifies coding up new matrices a lot.
Also added were precomputed matrices (which can be found in the script)
which significantly (tenfold) speed up computations. These could be
extended to the non-linear case using Taylor-expansions. It also negated
the need for the posterior class, which is still present in the code,
but actually redundant. These precomputed quantities are found in the
montecarlo::precomp_misfit() and montecarlo::precomp_misfitGrad()
functions.

First work was done on implementing generalised momentum. The sampling
is still unstable, for the proposed samples are uncorrelated. Drawing
an x amount of correlated samples from a multidimensional Gaussian is
an interesting task an sich, but this will not have priority. The plan
is to do it sometime using LU-decomposition.

More lines were replaced with actual operators instead of functions,
from the linearalgebra class.